Move slack endpoints to an independent file.
